Add omap drm display driver
A DSS based DRM display driver, which provides a plugin interface for
3d/2d accelerators to register. The core driver handles construction
of CRTC/encoder/connectors to represent the hardware, and support KMS.
This driver replaces omapfb. The omap drm driver allocates framebuffer
memory and implements (with the help of drm_fb_helper) the legacy fbdev
interface.

The driver maps CRTCs to overlays, encoders to overlay-managers, and
connectors to dssdev's.

Signed-off-by: Rob Clark <rob@ti.com>
Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>