syslink:ipu_pm: Modifications for timer setup for WDT
Modified to configure timer 3 that is used to trigger hibernation
and it is also used as watch dog timer. Since both ducati cores
can go to retention together and if one core is not responding
both need to be reset, we only use 1 timer

[Hari K] The free irq while freeing
checks the dev_id as the token with the dev_id that it was registered
earlier. To avoid the race condition as when the remote proc frees the
dmtimer handle save the handle in ipu_pm to use it as a token when passing
to free_irq
A cleaner solution should be investigated for watchdog timer

Change-Id: I15d3c0182d84861819cc36abf1f78cbc1674375f
Signed-off-by: Shahid Akhtar <sakhtar@ti.com>