omap:iodmm-cleanup and fixes
This patch handles the following:

1. cleaned the switch statement in dmm_user.c to move the copy_from_user
to individual functions.

2. Seperated the building of Scatter gather list to Map function. The list
is created when the Map function is created and deleted when the unmap
function is called. Previously this list was created during flush function
and deleted during the invalidate function. The previous implementation would
fail if the Users call the invalidate function directly without calling flush
function, as it would require calling flush function prior to calling invalidate
function. The previous implementation also includes the overhead in creating
and deleting the SG list for every flush and invalidate calls.

3. Changed the logic to track mapping object based on MPU VA instead of page
aligned MPU VA. This is because if there are 2 MPU buffers with virtual
addresses within the Page boundary, then there would be 2 Mapping objects with
same virtual address.

4. Delete the DMM pools only when all the handles to IODMM are closed. The
ioctl DMM_IOCDELETEPOOL can be deprecated with this change.

5. Added the SYS ADMIN check for create dmm pools, and set tlb entry ioctls.
This code is currently kept commented out.

6. DMA Flush APIs are still optionally selected using CONFIG_DMM_DMA_API option.
This is not selected by default.

Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>

omap:iodmm-send notification on close only when resources not free

Instead of sending the notification on every IOMMU close, send it
only when there are still mapped resources to avoid the remote
processor accessing these resources.
After compelting the notification, the iommu release function can
safely remove the mapped entries.

Change-Id: Ib162bb4aa4ec0ed287a4b7484f54b0712871804e
Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>