syslink: iommu: provide new api to return a device
The ipu device is missing from the omap_ipupm_data. This
results in a null device pointer being used when applying
constraints. This change provides an api which can be used
to return a device when queried by name. This is to support
initializing omap_ipupm_data with a device.

Change-Id: I490e61db6361652aa0fd63ba51e98af937eb800e
Signed-off-by: Bryan Honza <honza@ti.com>