TILER: Add support for in-page offset for buffers.
Now that blocks are not necessarily page-aligned, and the
physical addresses of blocks are not revealed, the buffer offset
has to contain the in-page offset.

This patch propagates the in-page offset into the buffer offset
and adjusts the buffer size so still only the actual pages are
used.

It also adjusts the offset lookup to be able to deal with
non-aligned buffer offsets and lengths.

Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: David Sin <davidsin@ti.com>