ARM: barrier: allow options to be passed to memory barrier instructions
BugLink: http://bugs.launchpad.net/bugs/1300343

commit 3ea128065ed20d33bd02ff6dab689f88e38000be upstream.

On ARMv7, the memory barrier instructions take an optional `option'
field which can be used to constrain the effects of a memory barrier
based on shareability and access type.

This patch allows the caller to pass these options if required, and
updates the smp_*() barriers to request inner-shareable barriers,
affecting only stores for the _wmb variant. wmb() is also changed to
use the -st version of dsb.

Reported-by: Albin Tonnerre <albin.tonnerre@arm.com>
Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
[ 3.8-stable prereq for
  7c8746a ARM: 7955/1: spinlock: ensure we have a compiler barrier before sev ]
Signed-off-by: Kamal Mostafa <kamal@canonical.com>
Signed-off-by: Joseph Salisbury <joseph.salisbury@canonical.com>